requirements.txt
certifi==2018.11.29
chardet==3.0.4
future==0.15.2
idna==2.7
oauthlib==1.1.2
pkg-resources==0.0.0
python-dateutil==2.8.1
python-twitter==3.5
requests==2.20.0
requests-oauthlib==0.6.2
six==1.13.0
urllib3==1.24.1
